Extract BucketConfig and BucketFSConnectionConfig from bucket_config.py
