+++ exasol_bucketfs_utils_python/download.py
@@ -6,7 +6,7 @@ import joblib
 import requests
 
 from exasol_bucketfs_utils_python import bucketfs_utils
-from exasol_bucketfs_utils_python.bucketfs_config import BucketConfig
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
 from exasol_bucketfs_utils_python.bucketfs_utils import generate_bucket_http_url
